requirements.txt
shiny
pandas
duckdb==1.3.2
ortools
python-dotenv
plotly
shinywidgets
